Rename settings panel components
`<PlaybackSettings/>` → `<BasicSettings/>`, `<VisualizationSettings/>` → `<AdvancedSettings/>`
