+++ src/components/AdvancedSettings.svelte
