+++ src/components/TabbedPanel.svelte
@@ -11,8 +11,8 @@
 <script>
   import PanelSwitcher from "./PanelSwitcher.svelte";
   import PlaybackControls from "./PlaybackControls.svelte";
-  import PlaybackSettings from "./PlaybackSettings.svelte";
-  import VisualizationSettings from "./VisualizationSettings.svelte";
+  import BasicSettings from "./BasicSettings.svelte";
+  import AdvancedSettings from "./AdvancedSettings.svelte";
 
   export let playPauseApp;
   export let skipToPercentage;
@@ -23,9 +23,9 @@
 
 <PanelSwitcher bind:selectedPanel />
 <div>
-  <PlaybackSettings {skipToPercentage} />
+  <BasicSettings {skipToPercentage} />
   {#if selectedPanel === 'settings'}
-    <VisualizationSettings />
+    <AdvancedSettings />
   {/if}
 </div>
 <PlaybackControls {playPauseApp} {stopApp} />
